backend/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
email-validator>=2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database - MongoDB
motor>=3.3.0
beanie>=1.25.0

# AI/ML - Semantic Grading
sentence-transformers>=2.2.2
torch>=2.2.0
numpy>=1.26.0

# OCR - Handwriting Recognition (with OpenCV preprocessing)
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
Pillow>=10.2.0

# Analytics
pandas>=2.1.0

# AI Question Generation
google-generativeai>=0.3.0

# Document Processing
# Note: PyPDF2 is deprecated — use pypdf (the maintained fork)
pypdf>=4.0.0
python-docx>=1.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
